# swagger copy.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Config Module",
    "license": {
      "name": "Proprietary",
      "url": "https://iotechglobal.io/licenses/LICENSE-1.0.html"
    },
    "x-vendor-info": {
      "name": "IoTech LLC",
      "url": "http://localhost:3000",
      "email": "admin@iotechglobal.io",
      "address": null,
      "street": null,
      "city": null,
      "state": null,
      "zip": null,
      "country": null
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
  "paths": {
...
    "/users": {
      "post": {
        "summary": "create user operation",
        "operationId": "createUser",
        "tags": [
          "User"
        ],
        "requestBody": {
          "description": "create user post request",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payload"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Ok",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "put": {
        "summary": "update user operation",
        "operationId": "updateUser",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "description": "update Customer",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Payload"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "delete user operation",
        "operationId": "deleteUser",
        "tags": [
          "User"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "description": "User ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Ok"
          },
          "404": {
            "description": "Resource not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          },
          "500": {
            "description": "Unexpected error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Payload"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "iotech_config_api": {
        "scheme": "bearer",
        "type": "http",
        "description": "Token based authentication"
      }
    },
    "schemas": {
      "Payload": {
        "allOf": [
          {
            "$ref": "#/components/schemas/Header"
          },
          {
            "type": "object",
            "required": [
              "data"
            ],
            "properties": {
              "data": {
                "$ref": "#/components/schemas/EntityDetails"
              },
              "result": {
                "$ref": "#/components/schemas/ResponseResult"
              },
              "error": {
                "$ref": "#/components/schemas/Error"
              },
              "stats": {
                "$ref": "#/components/schemas/Stats"
              },
              "metaData": {
                "$ref": "#/components/schemas/MetaData"
              }
            }
          }
        ]
      },
      "EntityDetails": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/Component"
          },
          {
            "$ref": "#/components/schemas/Measurement"
          },
          {
            "$ref": "#/components/schemas/Model"
          },
          {
            "$ref": "#/components/schemas/Asset"
          },
          {
            "$ref": "#/components/schemas/Image"
          },
          {
            "$ref": "#/components/schemas/Uom"
          },
          {
            "$ref": "#/components/schemas/Datasource"
          },
          {
            "$ref": "#/components/schemas/DatasourceNode"
          },
          {
            "$ref": "#/components/schemas/EntityDatasourceNodeLink"
          },
          {
            "$ref": "#/components/schemas/TagMappingLinks"
          },
          {
            "$ref": "#/components/schemas/EventConfig"
          },
          {
            "$ref": "#/components/schemas/Subscription"
          },
          {
            "$ref": "#/components/schemas/DsNodeTag"
          },
          {
            "$ref": "#/components/schemas/Perspective"
          },
          {
            "$ref": "#/components/schemas/PerspectiveTreeNode"
          },
          {
            "$ref": "#/components/schemas/Request"
          },
          {
            "$ref": "#/components/schemas/ComponentInstance"
          },
          {
            "$ref": "#/components/schemas/Property"
          },
          {
            "$ref": "#/components/schemas/Panel"
          },
          {
            "$ref": "#/components/schemas/Dashboard"
          },
          {
            "$ref": "#/components/schemas/AggConfig"
          },
          {
            "$ref": "#/components/schemas/AggDataRes"
          },
          {
            "$ref": "#/components/schemas/FetchAggDataReq"
          },
          {
            "$ref": "#/components/schemas/Annotation"
          },
          {
            "$ref": "#/components/schemas/writeback"
          },
          {
            "$ref": "#/components/schemas/AssetSummary"
          },
          {
            "$ref": "#/components/schemas/UserPerspectives"
          },
          {
            "$ref": "#/components/schemas/Customer"
          },
          {
            "$ref": "#/components/schemas/User"
          },
          {
            "$ref": "#/components/schemas/UserGroup"
          },
          {
            "$ref": "#/components/schemas/UserGroupMembers"
          },
          {
            "$ref": "#/components/schemas/ACLRequest"
          },
          {
            "$ref": "#/components/schemas/Feature"
          },
          {
            "$ref": "#/components/schemas/Role"
          },
          {
            "$ref": "#/components/schemas/RoleDelete"
          },
          {
            "$ref": "#/components/schemas/GetDsNode"
          },
          {
            "$ref": "#/components/schemas/GetAssetEventsRequest"
          },
          {
            "$ref": "#/components/schemas/RegisteredGroup"
          },
          {
            "$ref": "#/components/schemas/RefModelCompTypes"
          },
          {
            "$ref": "#/components/schemas/RefVendor"
          }
        ]
      },
      "Header": {
        "type": "object",
        "required": [
          "specVersion",
          "type",
          "source",
          "id",
          "time",
          "dataContentType"
        ],
        "properties": {
          "specVersion": {
            "type": "string",
            "example": "1.0.0"
          },
          "type": {
            "description": "Operation:(Get) EntityType: (TsRead)",
            "type": "string",
            "example": "<Operation>-<EntityType>-req"
          },
          "source": {
            "description": "The source system (client) initiating this request",
            "type": "string",
            "example": "intellija-ui"
          },
          "id": {
            "description": "Unique ID generated by the system that created this request",
            "type": "string",
            "example": "C234-1234-1234"
          },
          "time": {
            "description": "Time in ISO 8601 format when request created",
            "type": "string",
            "example": "2020-03-31T17:31:00.123Z"
          },
          "dataContentType": {
            "type": "string",
            "example": "application/json"
          }
        }
      },
      "Component": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "componentTypeId": {
            "type": "integer"
          },
          "vendorId": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "$ref": "#/components/schemas/Image"
          },
          "isActive": {
            "type": "boolean"
          },
          "sourceModelId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Property"
            }
          }
        }
      },
      "Measurement": {
        "type": "object",
        "required": [
          "name",
          "baseUom",
          "baseUomAbbreviation"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "derivedUom": {
            "type": "string"
          },
          "baseUom": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "baseUomAbbreviation": {
            "type": "string"
          }
        }
      },
      "Model": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "modelTypeId": {
            "type": "integer"
          },
          "notes": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "isActive": {
            "type": "boolean"
          },
          "image": {
            "$ref": "#/components/schemas/Image"
          },
          "sourceModelId": {
            "type": "string"
          },
          "components": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ComponentData"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "properties": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Property"
            }
          }
        }
      },
      "Asset": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "labels": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "image": {
            "$ref": "#/components/schemas/Image"
          },
          "isActive": {
            "type": "boolean"
          },
          "sourceModelId": {
            "type": "string"
          },
          "components": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "Image": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "data": {
            "type": "string",
            "maxLength": 600000
          }
        }
      },
...
      "Error": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32",
            "example": "500 or 400"
          },
          "message": {
            "type": "string",
            "example": "something went wrong"
          }
        }
      },
      "ResponseResult": {
        "type": "object",
        "required": [
          "corrId",
          "status"
        ],
        "properties": {
          "status": {
            "type": "string",
            "example": "ok"
          },
          "corrId": {
            "type": "string",
            "example": "C234-1234-1234",
            "description": "The value in this field should be the same as what was received in the \"id\" field of the corresponding request"
          }
        }
      },
      "Stats": {
        "type": "object",
        "required": [
          "requestReceivedTime",
          "dbExecTimeMs",
          "totalExecTimeMs"
        ],
        "properties": {
          "requestReceivedTime": {
            "description": "Time when the server received this request. Note that it contains the millseconds part too.",
            "type": "string",
            "example": "2020-03-31T17:31:14.235Z"
          },
          "dbExecTimeMs": {
            "description": "Time it took (in milliseconds) for the overall request to be processed by the db server. Blank in error case",
            "type": "number",
            "format": "double"
          },
          "totalExecTimeMs": {
            "description": "Time it took (in milliseconds) for the overall request to be processed by the server",
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "MetaData": {
        "type": "object",
        "required": [
          "moreRecordsAvailable",
          "nextRecordOffset",
          "recordsCount"
        ],
        "properties": {
          "moreRecordsAvailable": {
            "description": "if more records available on that request for next page",
            "type": "string",
            "example": "2020-03-31T17:31:14.235Z"
          },
          "nextRecordOffset": {
            "description": "next offset number for fetching next page records",
            "type": "integer"
          },
          "recordsCount": {
            "description": "total number of records in response after applying all filters requested by user",
            "type": "integer"
          }
        }
      },
      "Datasource": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Auto generated UUID",
            "type": "string"
          },
          "name": {
            "description": "Datasource name",
            "type": "string"
          },
          "datasourceTypeId": {
            "description": "reference UUID of datasource types",
            "type": "string"
          },
          "config": {
            "description": "Configuration properties in JSON format",
            "type": "object"
          }
        }
      },
      "DatasourceNode": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Auto generated UUID",
            "type": "string"
          },
          "name": {
            "description": "DS Node name",
            "type": "string"
          },
          "type": {
            "description": "reference UUID of DS Node Type",
            "type": "string"
          },
          "datasourceId": {
            "description": "reference UUID of Datasource",
            "type": "string"
          },
          "config": {
            "description": "Configuration properties in JSON format",
            "type": "object"
          }
        }
      },
      "Uom": {
        "type": "object",
        "required": [
          "name",
          "factor",
          "offset",
          "symbol",
          "measurementId"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "symbol": {
            "type": "string"
          },
          "baseUomId": {
            "type": "integer"
          },
          "measurementId": {
            "type": "integer"
          },
          "uomGroupId": {
            "type": "integer"
          },
          "factor": {
            "type": "number"
          },
          "offset": {
            "type": "number"
          },
          "referenceUomId": {
            "type": "integer"
          },
          "referenceFactor": {
            "type": "number"
          },
          "referenceOffset": {
            "type": "number"
          },
          "isBaseUom": {
            "type": "boolean"
          }
        }
      },
      "EntityDatasourceNodeLink": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Auto generated UUID",
            "type": "string"
          },
          "entityId": {
            "description": "Entity ID",
            "type": "string"
          },
          "datasourceNodeId": {
            "description": "Datasource Node Id",
            "type": "string"
          },
          "effectiveFrom": {
            "description": "date effective from",
            "type": "string"
          },
          "effectiveTo": {
            "description": "date effective To",
            "type": "string"
          },
          "isDeleted": {
            "description": "is deleted flag",
            "type": "boolean"
          },
          "isLinked": {
            "description": "is Linked flag",
            "type": "boolean"
          }
        }
      },
      "TagMappingLinks": {
        "type": "object",
        "properties": {
          "commentCode": {
            "type": "string"
          },
          "linkList": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TagMappingLink"
            }
          }
        }
      },
...
      "ComponentInstance": {
        "type": "object",
        "required": [
          "id",
          "name",
          "modelId",
          "componentId"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "modelId": {
            "type": "string"
          },
          "componentId": {
            "type": "string"
          }
        }
      },
      "EventConfig": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "eventConfigType": {
            "type": "string"
          },
          "parentConfigId": {
            "type": "string"
          },
          "isEnabled": {
            "type": "boolean"
          },
          "isOverride": {
            "type": "boolean"
          },
          "eventConfigDetails": {
            "$ref": "#/components/schemas/EventConfigDetails"
          },
          "eventActions": {
            "type": "object"
          },
          "additionalProperties": {
            "type": "object"
          }
        }
      },
...
      "Subscription": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string"
          },
          "eventTypeId": {
            "type": "string"
          },
          "subscriptionDetails": {
            "$ref": "#/components/schemas/SubscriptionDetails"
          }
        }
      },
...
      "DsNodeTag": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Auto generated UUID",
            "type": "string"
          },
          "name": {
            "description": "DS Node Tag name",
            "type": "string"
          },
          "dsNodeId": {
            "description": "reference UUID of DS Node ",
            "type": "string"
          },
          "config": {
            "description": "Configuration properties in JSON format",
            "type": "object"
          },
          "dsUom": {
            "description": "Unit of measurement",
            "type": "string"
          }
        }
      },
      "Perspective": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Autogenerated UUID",
            "type": "string"
          },
          "name": {
            "description": "Perspective Name",
            "type": "string"
          },
          "displayName": {
            "description": "Perspective Display Name",
            "type": "string"
          },
          "isDefault": {
            "description": "Only used for get perspective ofr user",
            "type": "boolean"
          },
          "userGroups": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "PerspectiveTreeNode": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Autogenerated UUID",
            "type": "string"
          },
          "perspectiveId": {
            "description": "Perspective Id",
            "type": "string"
          },
          "parentId": {
            "description": "refers to id of same entity type as parent relation",
            "type": "string"
          },
          "nodeIndex": {
            "description": "Node index",
            "type": "integer"
          },
          "name": {
            "description": "Perspective Tree Node Name",
            "type": "string"
          },
          "treeNodeType": {
            "description": "Tree Node Type",
            "type": "string"
          },
          "entityId": {
            "description": "Asset Id",
            "type": "string"
          },
          "latitude": {
            "description": "Latitude",
            "type": "number"
          },
          "longitude": {
            "description": "Longitude",
            "type": "number"
          }
        }
      },
      "Request": {
        "type": "object",
        "required": [
          "username",
          "password"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
...
      "Property": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "value": {
            "type": "string"
          },
          "dataTypeId": {
            "type": "string"
          },
          "assetId": {
            "type": "string"
          },
          "componentInstanceId": {
            "type": "string"
          },
          "parentPropertyId": {
            "type": "string"
          },
          "isOverride": {
            "type": "boolean"
          }
        }
      },
      "Panel": {
        "type": "object",
        "required": [
          "name",
          "themeId",
          "entityId",
          "perspectiveGroupId",
          "chartId",
          "rows",
          "cols",
          "panelType"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "themeId": {
            "type": "number"
          },
          "name": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "perspectiveGroupId": {
            "type": "string"
          },
          "chartId": {
            "type": "number"
          },
          "rows": {
            "type": "number"
          },
          "cols": {
            "type": "number"
          },
          "config": {
            "type": "object"
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/panelTag"
            }
          },
          "aggConfigs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "panelType": {
            "type": "string"
          }
        }
      },
...
      "Dashboard": {
        "type": "object",
        "required": [
          "name",
          "entityId",
          "panels",
          "path"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "perspectiveGroupId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "isDefault": {
            "type": "boolean"
          },
          "isAggDashboard": {
            "type": "boolean"
          },
          "panels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/dashboardPanelObj"
            }
          }
        }
      },
...
      "AggConfig": {
        "type": "object",
        "required": [
          "funcId",
          "tags",
          "name",
          "type",
          "aggrConfigs"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "funcId": {
            "type": "integer"
          },
          "perspectiveGroupId": {
            "type": "string"
          },
          "entityId": {
            "type": "string"
          },
          "timeRangeId": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object",
              "$ref": "#/components/schemas/aggConfigTag"
            }
          },
          "aggrConfigs": {
            "type": "array",
            "items": {
              "type": "object",
              "$ref": "#/components/schemas/FetchAggConfigObj"
            }
          },
          "funcConfig": {
            "$ref": "#/components/schemas/funcConfig"
          }
        }
      },
...
      "AggDataRes": {
        "type": "object",
        "properties": {
          "configs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AggConfigView"
            }
          }
        }
      },
      "FetchAggDataReq": {
        "type": "object",
        "properties": {
          "aggConfigs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/FetchAggConfigObj"
            }
          }
        }
      },
...
      "Annotation": {
        "type": "object",
        "required": [
          "name",
          "assetId",
          "timestamp",
          "tags"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "assetId": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "description": "format should be ' yyyy-MM-dd'T'HH:mm:ss.SSS'Z' '"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "object"
            }
          }
        }
      },
      "writeback": {
        "type": "object",
        "required": [
          "tags",
          "authDetails"
        ],
        "properties": {
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WriteBackTag"
            }
          },
          "authDetails": {
            "$ref": "#/components/schemas/Request"
          }
        }
      },
...
      "AssetSummary": {
        "type": "object",
        "properties": {
          "isLinked": {
            "type": "boolean",
            "description": "return true if any ds node with atleast 1 tag is linked"
          },
          "lastUnlinkTime": {
            "type": "string",
            "description": "return timestamp of last Ds Node unlink, if isLinked is false"
          },
          "hasAlarm": {
            "type": "boolean",
            "description": "return true if any alarm is active on that Asset"
          },
          "lastTagUpdatedTime": {
            "type": "string",
            "description": "return timestamp of any tag that was updated with liveData recently."
          },
          "linkedTags": {
            "type": "array",
            "items": {
              "type": "object"
            },
            "description": "if isLinked is true, this list with contains all tags which are linked. each entry will have comp instance id, tag id and timestamp when it was linked"
          }
        }
      },
      "UserPerspectives": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/Perspective"
        }
      },
      "Customer": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "username",
          "firstName",
          "lastName",
          "email",
          "mobileNo",
          "customerId"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "username": {
            "type": "string"
          },
          "firstName": {
            "type": "string"
          },
          "middleName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "mobileNo": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "customerId": {
            "type": "string"
          }
        }
      },
      "UserGroup": {
        "type": "object",
        "required": [
          "name",
          "path"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "path": {
            "type": "string"
          },
          "roles": {
            "type": "array"
          }
        }
      },
      "Feature": {
        "type": "object",
        "required": [
          "name",
          "displayName"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          }
        }
      },
      "UserGroupMembers": {
        "type": "object",
        "required": [
          "users",
          "groups"
        ],
        "properties": {
          "users": {
            "type": "array"
          },
          "groups": {
            "type": "array"
          }
        }
      },
      "ACLRequest": {
        "type": "object",
        "required": [
          "acls"
        ],
        "properties": {
          "acls": {
            "type": "array",
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ACLsCreate"
                },
                {
                  "$ref": "#/components/schemas/ACLsUpdate"
                },
                {
                  "$ref": "#/components/schemas/ACLsDelete"
                }
              ]
            }
          }
        }
      },
...
      "RoleDelete": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "name": {
            "type": "string"
          }
        }
      },
...
    }
  }
}